requirements.txt
Flask
firebase-admin
Gunicorn
segno           
Pillow
time